+++ packages/react-app/src/actions/index.js
@@ -1,2 +1,2 @@
 export { default as userReducer, setUserAddress } from "./userSlice";
-export { default as networkReducer, setExchangePrice } from "./networkSlice";
+export { default as networkReducer, setExchangePrice, setGasPrice } from "./networkSlice";
