+++ packages/react-app/src/actions/networkSlice.js
@@ -4,14 +4,18 @@ export const networkSlice = createSlice({
   name: "network",
   initialState: {
     exchangePrice: null,
+    gasPrice: null,
   },
   reducers: {
     setExchangePrice: (state, price) => {
       state.exchangePrice = price.payload;
     },
+    setGasPrice: (state, price) => {
+      state.price = price.payload;
+    },
   },
 });
 
-export const { setExchangePrice } = networkSlice.actions;
+export const { setExchangePrice, setGasPrice } = networkSlice.actions;
 
 export default networkSlice.reducer;
