+++ packages/react-app/src/hooks/UpdateParcels.js
@@ -1,15 +1,15 @@
 import { fetchParcelMetadata } from "../data";
 
-const useUpdateParcels = async (currentParcels, setParcels, contractToRead, forceUpdate = false) => {
+const useUpdateParcels = async (currentParcels, setParcels, readContracts, forceUpdate = false) => {
   var newParcels = [];
   if (currentParcels.length > 0 && !forceUpdate) return; // prevent excessive calls to IPFS
-  if (contractToRead) {
-    const parcelIds = await contractToRead.CityDaoParcel.getParcelIds();
-    const parcelURIs = await contractToRead.CityDaoParcel.getListedParcels();
+  if (readContracts) {
+    const parcelIds = await readContracts.CityDaoParcel.getParcelIds();
+    const parcelURIs = await readContracts.CityDaoParcel.getListedParcels();
     for (var index = 0; index < parcelIds.length; index++) {
       const parcelId = parcelIds[index];
       const ipfsHash = parcelURIs[parcelId];
-      const price = await contractToRead.CityDaoParcel.getPrice(parcelId);
+      const price = await readContracts.CityDaoParcel.getPrice(parcelId);
       if (ipfsHash !== "") {
         // skip sold parcels
         try {
